requirements.txt
Django==1.7.7
hamlpy
django-nose
model_mommy
django-bootstrap3
http://darcs.idyll.org/~t/projects/figleaf-0.6.1.tar.gz
specloud
Pillow
should-dsl
django-blog-zinnia
django-tinymce
zinnia-wysiwyg-tinymce
django-app-namespace-template-loader
zinnia-theme-bootstrap
django-toolbelt
django-grappelli
sorl-thumbnail
django-bootstrap-pagination